requirements.txt
paho-mqtt~=1.6.1
Flask~=2.3.2
requests~=2.31.0